requirements.txt
setuptools_scm
pillow
python-dotenv
requests
ujson
tinydb